+++ canister/src/http/mod.rs
@@ -1,13 +1,12 @@
 pub mod errors;
 
-use crate::metrics::MetricRpcCallResponse;
 use crate::{
     add_latency_metric, add_metric_entry,
     constants::{COLLATERAL_CYCLES_PER_NODE, CONTENT_TYPE_VALUE},
     http::errors::HttpClientError,
     logs::Priority,
     memory::{next_request_id, read_state, State},
-    metrics::{MetricRpcHost, MetricRpcMethod},
+    metrics::{MetricRpcCallResponse, MetricRpcHost, MetricRpcMethod},
 };
 use canhttp::{
     convert::ConvertRequestLayer,
@@ -70,11 +69,6 @@ where
                         request_id: req.body().id().clone(),
                         start_ns: ic_cdk::api::time()
                     };
-                    add_metric_entry!(
-                        requests,
-                        (req_data.method.clone(), req_data.host.clone()),
-                        1
-                    );
                     log!(Priority::TraceHttp, "JSON-RPC request with id `{}` to {}: {:?}",
                         req_data.request_id,
                         req_data.host.0,
@@ -195,6 +189,11 @@ fn observe_response(response: MetricRpcCallResponse, req_data: &MetricData) {
             // Don't record latency for IC errors
         }
     }
+    add_metric_entry!(
+        requests,
+        (req_data.method.clone(), req_data.host.clone()),
+        1
+    );
     add_metric_entry!(
         responses,
         (req_data.method.clone(), req_data.host.clone(), response),